server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
composio-core==0.6.15
composio-langchain==0.6.15
composio-openai==0.6.15
cryptography==44.0.0
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
inflection==0.5.1
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-core==0.3.30
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langchainhub==0.1.21
langsmith==0.2.11
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.4
openai==1.59.8
orjson==3.10.15
packaging==24.2
paramiko==3.5.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
pygments==2.19.1
pynacl==1.5.0
pyperclip==1.9.0
pysher==1.0.8
python-dotenv==1.0.1
pyyaml==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
semver==3.0.3
sentry-sdk==2.20.0
sniffio==1.3.1
sqlalchemy==2.0.37
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
yarl==1.18.3
zipp==3.21.0
